Refactor PageDetails; Now renders in AppDetails with side drawer and header
